third_party/WebKit/LayoutTests/fast/writing-mode/vertical-inline-block-hittest.html

<!DOCTYPE html>
<html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body style="margin: 0px">
<div id="target" style="-webkit-writing-mode: vertical-rl; border: solid 1px black; font-family: Ahem">
<br />
<span style="display: inline-block"><a>b</a></span>
</div>
<script>
test(() => {
    elementInsideInlineBlockInSecondLine = document.querySelector("a");
    assert_equals(document.elementFromPoint(5, 5), elementInsideInlineBlockInSecondLine);
}, 'Click 2nd line inside inline-block in vertical writing mode');
</script>
</body>
</html>
